Made an "Alleyway" folder to keep all of the art assets for this level
The stage will start with "Alleyway pattern start" and either continue with the "cont" version or the individual buildings to build your own length/pattern with. The electric lines and the tall lights are just there for detail. Level design team, feel free to move all of these stuff around to fit the level better.